ggtg

+++ Service.Host/Service.Host/Services/Test.cs
@@ -80,20 +80,26 @@ namespace Service.Host.Services
                         Console.WriteLine(threadsWrapper[num].service.DeleteAccountingObject(generator.GenerateInventaryNumber(7)));
                         break;
                     case 3:
-                        Console.WriteLine("===========================================");
-                        inventaryNumber = objectList.Items[rnd.Next(objectList.ItemsCount)].GetValue("InventaryNumber").ToString();
-                        Console.WriteLine("Удаляю объект учета с инв. номером {0}", inventaryNumber);
-                        Console.WriteLine();
-                        Console.WriteLine(threadsWrapper[num].service.DeleteAccountingObject(inventaryNumber));
+                        if (objectList.ItemsCount != 0)
+                        {
+                            Console.WriteLine("===========================================");
+                            inventaryNumber = objectList.Items[rnd.Next(objectList.ItemsCount)].GetValue("InventaryNumber").ToString();
+                            Console.WriteLine("Удаляю объект учета с инв. номером {0}", inventaryNumber);
+                            Console.WriteLine();
+                            Console.WriteLine(threadsWrapper[num].service.DeleteAccountingObject(inventaryNumber));
+                        }
                         break;
                     case 4:
-                        Console.WriteLine("===========================================");
-                        oldInventaryNumber = objectList.Items[rnd.Next(objectList.ItemsCount)].GetValue("InventaryNumber").ToString();
-                        Console.WriteLine(String.Format("Изменяю объект учета (инв.номер: {0}.)\nИнв. номер: {1}\nОписание: {2}\nДата оприходования: {3}\nДата амортизации: {4}\nМОЛ: {5}",
-                        oldInventaryNumber, inventaryNumber, description, postingDate, deprecationDate, owner));
-                        Console.WriteLine();
-                        Console.WriteLine(threadsWrapper[num].service.UpdateAccountingObject(oldInventaryNumber, inventaryNumber,
-                            description, postingDate, deprecationDate, owner));
+                        if (objectList.ItemsCount != 0)
+                        {
+                            Console.WriteLine("===========================================");
+                            oldInventaryNumber = objectList.Items[rnd.Next(objectList.ItemsCount)].GetValue("InventaryNumber").ToString();
+                            Console.WriteLine(String.Format("Изменяю объект учета (инв.номер: {0}.)\nИнв. номер: {1}\nОписание: {2}\nДата оприходования: {3}\nДата амортизации: {4}\nМОЛ: {5}",
+                            oldInventaryNumber, inventaryNumber, description, postingDate, deprecationDate, owner));
+                            Console.WriteLine();
+                            Console.WriteLine(threadsWrapper[num].service.UpdateAccountingObject(oldInventaryNumber, inventaryNumber,
+                                description, postingDate, deprecationDate, owner));
+                        }
                         break;
 
                 }
@@ -104,19 +110,12 @@ namespace Service.Host.Services
         public void Run()
         {
             Random rnd = new Random();
-            threadsWrapper.Add(new ThreadModel());
-            threadsWrapper[threadsWrapper.Count() - 1].thread = new Thread(TestItem);
-            threadsWrapper[threadsWrapper.Count() - 1].thread.Name = (threadsWrapper.Count() - 1).ToString();
-            threadsWrapper[threadsWrapper.Count() - 1].thread.Start();
-            while (true)
+            for (int i = 0; i < maxServicesCount; ++i)
             {
-                if ((threadsWrapper.Count() < maxServicesCount) && (rnd.Next() < 0.4))
-                {
-                    threadsWrapper.Add(new ThreadModel());
-                    threadsWrapper[threadsWrapper.Count() - 1].thread = new Thread(TestItem);
-                    threadsWrapper[threadsWrapper.Count() - 1].thread.Name = (threadsWrapper.Count() - 1).ToString();
-                    threadsWrapper[threadsWrapper.Count() - 1].thread.Start();
-                }
+                threadsWrapper.Add(new ThreadModel());
+                threadsWrapper[threadsWrapper.Count() - 1].thread = new Thread(TestItem);
+                threadsWrapper[threadsWrapper.Count() - 1].thread.Name = (threadsWrapper.Count() - 1).ToString();
+                threadsWrapper[threadsWrapper.Count() - 1].thread.Start();
             }
         }
